JAMES-3088 Favor composition for UsersRepository
Rely on a DAO rather than on inheritance.

Note that we need API stability for JPA, memory & LDAP for Spring product
