+++ server/data/data-cassandra/src/test/java/org/apache/james/user/cassandra/CassandraUsersRepositoryTest.java
@@ -22,8 +22,8 @@ package org.apache.james.user.cassandra;
 import org.apache.commons.configuration2.BaseHierarchicalConfiguration;
 import org.apache.james.backends.cassandra.CassandraClusterExtension;
 import org.apache.james.domainlist.api.DomainList;
-import org.apache.james.user.lib.AbstractUsersRepository;
-import org.apache.james.user.lib.AbstractUsersRepositoryContract;
+import org.apache.james.user.lib.UsersRepositoryImpl;
+import org.apache.james.user.lib.UsersRepositoryContract;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -34,11 +34,11 @@ class CassandraUsersRepositoryTest {
     static CassandraClusterExtension cassandraCluster = new CassandraClusterExtension(CassandraUsersRepositoryModule.MODULE);
 
     @Nested
-    class WhenEnableVirtualHosting implements AbstractUsersRepositoryContract.WithVirtualHostingContract {
+    class WhenEnableVirtualHosting implements UsersRepositoryContract.WithVirtualHostingContract {
         @RegisterExtension
         UserRepositoryExtension extension = UserRepositoryExtension.withVirtualHost();
 
-        private CassandraUsersRepository usersRepository;
+        private UsersRepositoryImpl usersRepository;
 
         @BeforeEach
         void setUp(TestSystem testSystem) throws Exception {
@@ -46,17 +46,17 @@ class CassandraUsersRepositoryTest {
         }
 
         @Override
-        public AbstractUsersRepository testee() {
+        public UsersRepositoryImpl testee() {
             return usersRepository;
         }
     }
 
     @Nested
-    class WhenDisableVirtualHosting implements AbstractUsersRepositoryContract.WithOutVirtualHostingContract {
+    class WhenDisableVirtualHosting implements UsersRepositoryContract.WithOutVirtualHostingContract {
         @RegisterExtension
         UserRepositoryExtension extension = UserRepositoryExtension.withoutVirtualHosting();
 
-        private CassandraUsersRepository usersRepository;
+        private UsersRepositoryImpl usersRepository;
 
         @BeforeEach
         void setUp(TestSystem testSystem) throws Exception {
@@ -64,16 +64,18 @@ class CassandraUsersRepositoryTest {
         }
 
         @Override
-        public AbstractUsersRepository testee() {
+        public UsersRepositoryImpl testee() {
             return usersRepository;
         }
     }
 
-    private static CassandraUsersRepository getUsersRepository(DomainList domainList, boolean enableVirtualHosting) throws Exception {
-        CassandraUsersRepository cassandraUsersRepository = new CassandraUsersRepository(domainList, cassandraCluster.getCassandraCluster().getConf());
+    private static UsersRepositoryImpl getUsersRepository(DomainList domainList, boolean enableVirtualHosting) throws Exception {
+        CassandraUsersDAO usersDAO = new CassandraUsersDAO(cassandraCluster.getCassandraCluster().getConf());
         BaseHierarchicalConfiguration configuration = new BaseHierarchicalConfiguration();
         configuration.addProperty("enableVirtualHosting", String.valueOf(enableVirtualHosting));
-        cassandraUsersRepository.configure(configuration);
-        return cassandraUsersRepository;
+
+        UsersRepositoryImpl usersRepository = new UsersRepositoryImpl(domainList, usersDAO);
+        usersRepository.configure(configuration);
+        return usersRepository;
     }
 }
